+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @graphtrace/web
 
+## 1.8.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @graphtrace/shared@1.8.0
+
 ## 1.7.2
 
 ### Patch Changes